% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfabs.r
\name{qfabs}
\alias{qfabs}
\title{A minorization-maximization forward and backward stagewise algorithm for high-dimensional quantile regression.}
\usage{
    qfabs(y, x, tau = 0.5, epsilon = 0.01, delta = 1e-8, xi = 1e-10, max.iter = 5000, gamma = 1)
}
\arguments{
    \item{y}{The response, a numeric vector.}
    \item{x}{The design matrix.}
    \item{tau}{The quantile to be estimated, this is a number strictly between 0 and 1. Default value is 0.5.}
    \item{epsilon}{The step size for updating coefficients. Default is 0.01.}
    \item{delta}{A constant in pertubed loss function. Default is 1e-8.}
    \item{xi}{The threshold for qfabs. Default is 1e-10.}
    \item{max.iter}{The maximum number of outer-loop iterations allowed. Default is 5000.}
    \item{gamma}{A tuning parameter in EBIC. Default is 1.}
}
\value{
A list.
\itemize{
    \item beta - The optimal solution that minimizes EBIC.
    \item lambda - lambda sequence generated by qfabs.
    \item iter - Iterations.
    \item EBIC - The EBIC for each solution.
    \item opt - Position of the optimal lambda based on EBIC.
}
}
\examples{
library("mnormt")
n = 200
p = 500
beta = matrix(rep(0, p), ncol = 1)
index = 1:5
beta[index, 1] = c(1.0,1.1,1.2,1.3,1.4)
distr = "t3"
rho = 0.8
dat = generator(n, p, beta, distr, rho)
x = dat$x
y = dat$y
tau = 0.2
fit <- qfabs(y, x, tau)
}
