% Generated by roxygen2: do not edit by hand
% Please edit documentation in data/
\name{simulatedData}
\docType{data}
\title{Simulated datasets from linear models.}
\usage{
    data(simulatedData_gaussian)
}

\description{
    Simulated data from linear models, including family 'gaussian' (simulatedData_gaussian), 't3' (simulatedData_t3), 'cauchy' (simulatedData_cauchy), and 'logistic' (simulatedData_logistic).
}
\examples{
    data(simulatedData_gaussian)
    x = data_gaussian$x
    y = data_gaussian$y
    n = data_gaussian$n
}
